requirements.txt
altair==5.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
azure-functions==1.11.2
azure-iot-device==2.13.0
azure-iot-hub==2.6.1
azure-storage-blob==12.20.0
blinker==1.7.0
branca==0.7.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
fastapi==0.110.1
uvicorn==0.29.0
azure-storage-blob==12.20.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-monitor-opentelemetry==1.6.0
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
fiona==1.9.5
flasgger==0.9.7.1
Flask==3.0.2
flask-sqlacodegen==2.0.0
Flask-SQLAlchemy==3.1.1
folium==0.14.0
fonttools==4.49.0
geopandas==0.14.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
hydra-core==1.3.2
hydra_zen==0.12.1
idna==3.6
importlib-metadata==6.11.0
inflect==7.0.0
iniconfig==2.0.0
ipyleaflet==0.18.2
ipython==8.21.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
numpy==1.25.1
omegaconf==2.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
Pillow==9.5.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==1.10.14
pydeck==0.8.1b0
pyodbc==5.1.0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pymssql==2.2.11
pyparsing==3.1.1
pypng==0.20220715.0
pyproj==3.6.1
pystac==1.6.1
pytest==8.0.2
python-barcode==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
shapely==2.0.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.27
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
streamlit==1.25.0
streamlit-folium==0.14.0
StrEnum==0.4.15
supervisor==4.2.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
traittypes==0.2.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==1.26.5
validators==0.22.0
rasterio==1.3.9
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
xyzservices==2023.10.1
zipp==3.17.0
